refactor(kitchenpos/commons): rename common package to commons
