+++ src/test/java/kitchenpos/menus/tobe/domain/Fixtures.java
@@ -1,8 +1,8 @@
 package kitchenpos.menus.tobe.domain;
 
-import kitchenpos.common.tobe.domain.DisplayedName;
-import kitchenpos.common.tobe.domain.Price;
-import kitchenpos.common.tobe.domain.Quantity;
+import kitchenpos.commons.tobe.domain.DisplayedName;
+import kitchenpos.commons.tobe.domain.Price;
+import kitchenpos.commons.tobe.domain.Quantity;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
