+++ src/test/java/kitchenpos/products/tobe/domain/ProductTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.products.tobe.domain;
 
-import kitchenpos.common.tobe.domain.Price;
+import kitchenpos.commons.tobe.domain.Price;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
